ObliQ/requirements.txt
numpy>=1.26
scipy>=1.11
sympy>=1.12
perceval-quandela==0.11.0
tqdm>=4.66
